refactor checks part.4
